[*] - save debts added

+++ Debt Manager/Debt Manager/Classes/Controllers/DMMainViewController.m
@@ -55,8 +55,8 @@
         NSInteger currentBorrowed = [[DataManager sharedInstance] amountForType:BORROW_TYPE];
         NSInteger currentLent     = [[DataManager sharedInstance] amountForType:LEND_TYPE];
         
-        self.borrowBalanceLabel.text = [NSString stringWithFormat:@"%ld/%ld",currentBorrowed ,(long)maxBorrowed];
-        self.borrowBalanceLabel.text = [NSString stringWithFormat:@"%ld/%ld",currentLent, (long)maxLent];
+        self.borrowBalanceLabel.text = [NSString stringWithFormat:@"%ld/%ld",(long)currentBorrowed ,(long)maxBorrowed];
+        self.borrowBalanceLabel.text = [NSString stringWithFormat:@"%ld/%ld",(long)currentLent, (long)maxLent];
         
         if (currentBorrowed > maxBorrowed) {
             self.borrowBalanceLabel.textColor = [UIColor redColor];
